BRUCE-USB-GUI/USBGUI/requirements.txt
pyserial
Pillow
platformio
